[Clean] del redigo
