+++ src/core/glide.yaml
@@ -2,7 +2,4 @@ package: .
 import:
 - package: github.com/go-xorm/xorm
 - package: github.com/golang/glog
-- package: github.com/garyburd/redigo
-  subpackages:
-  - redis
 - package: gopkg.in/redis.v5
